// src/api/gen/models/MediaSearchRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { MediaFilter } from "./MediaFilter";
/**
 * Запрос на поиск медиа
 */
export type MediaSearchRequest = {
  filter?: MediaFilter;
  page?: number;
  size?: number;
};